feat: add homework 20 project (Spring MVC)

+++ home_work_3/demo-application/src/main/java/ru.mts.demofintech/repository/AnimalRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.mts.demofintech.entity.Animal;
 
 @Repository
-public interface AnimalRepository extends JpaRepository<Animal, Integer> {
+public interface AnimalRepository extends JpaRepository<Animal, Long> {
 }
